feat: Modificar o endpoint DELETE /api/posts/{id} para remover um post do banco

+++ src/main/java/com/compass/microservicoB/controller/PostController.java
@@ -52,7 +52,12 @@ public class PostController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePost(@PathVariable String id) {
-        postService.deletePost(id);
-        return ResponseEntity.noContent().build();
+        Optional<Post> post = postService.getPostById(id);
+        if (post.isPresent()) {
+            postService.deletePost(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
